+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # Change Log
 
+[BACK TO MAIN README](README.md)
+
 All notable changes to the "f5-fast" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [0.0.1] - (4-24-2020)
 
-- Initial release
+- Initial release!
+    - Make api to get Chuck Norris joke, display in new editor window for testing, learning and laughs...
+    - Started extension settings to host devices
+    - Command skeleton for next call to be coded